refactor(types): параметры и команды формы вынесены в свой регистратор
Третий, последний разрез `FormTypesProvider`: `FormParametersRegistrar` забрал
структуру параметров управляемой формы, параметры-свойства обычной (вместе с
подстановкой владельца и подавлением несуществующих) и команды. Провайдер —
1820 → 730 строк и 84 → 32 метода, то есть ниже порогов `java:S1448` и
`java:S1200`, из-за которых резали.
